backend/requirements.txt
# Backend dependencies
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
cachetools==5.3.1
gunicorn==21.2.0
pytest==7.4.0
